♻️ refactor: mock api 코드를 api/ 폴더로 이동
